% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-get-functions.R
\name{get_lang}
\alias{get_lang}
\title{Get Language}
\usage{
get_lang()
}
\description{
Get Language
}
\seealso{
Other set-get-functions: \code{\link{get_year}};
  \code{\link{set_lang}}; \code{\link{set_year}}
}

