% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-get-functions.R
\name{get_year}
\alias{get_year}
\title{Get year}
\usage{
get_year()
}
\description{
Get year
}
\seealso{
Other set-get-functions: \code{\link{get_lang}};
  \code{\link{set_lang}}; \code{\link{set_year}}
}

