% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ghostscript_path}
\alias{ghostscript_path}
\title{Returns the path to the Ghostscript executable}
\usage{
ghostscript_path()
}
\description{
Returns the path to the Ghostscript executable bundled with the package on Windows, on other systems it returns the value of \code{Sys.which("gs")}.
}

