% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdf.to.png.R
\name{pdf.to.png}
\alias{pdf.to.png}
\title{Convert PDF to PNG}
\usage{
pdf.to.png(file, outfile, fontpaths = system.file("fonts", package =
  "pmbundle"), resolution = 600, options = "")
}
\arguments{
\item{file}{a character string giving the name of the original pdf file}

\item{outfile}{the name of the new png file including ".png" extension}

\item{fontpaths}{a character vector giving directories that Ghostscript will search for fonts}

\item{resolution}{target resolution of the png file, default is 600}

\item{options}{a character string containing further options to Ghostscript}
}
\description{
Deprecated, use \code{format.plot::pdf_to_png} instead. Converts a PDf file to PNG with Ghostscript
}

