% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.r
\name{percent_format}
\alias{percent_format}
\title{Percent format}
\usage{
percent_format()
}
\description{
Convert number to percent format for labels in ggplot
}
\seealso{
Other format: \code{\link{precision}}
}

