% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colors.r
\name{pm_colors_cmyk}
\alias{pm_colormap}
\alias{pm_colors}
\alias{pm_colors_cmyk}
\alias{pm_colors_rgb}
\title{Get pm colors}
\usage{
pm_colors_cmyk()

pm_colors_rgb()

pm_colors()

pm_colormap()
}
\description{
Functions to get color series

Functions to get color series

Plot pm colors
}
\seealso{
Other color functions: \code{\link{cmyk2hex}},
  \code{\link{cmyk2rgb}}, \code{\link{rgb2cmyk}}

Other color functions: \code{\link{cmyk2hex}},
  \code{\link{cmyk2rgb}}, \code{\link{rgb2cmyk}}

Other color functions: \code{\link{cmyk2hex}},
  \code{\link{cmyk2rgb}}, \code{\link{rgb2cmyk}}
}

