% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.r
\name{Rmd2md}
\alias{Rmd2md}
\title{Convert Rmd to Markdown}
\usage{
Rmd2md(input, dir.path, embed.fonts = TRUE, output.tex = "",
  replace.backticks = TRUE, ...)
}
\arguments{
\item{input}{input file}

\item{dir.path}{file path}

\item{embed.fonts}{embed fonts (default: TRUE)}

\item{output.tex}{output tex}

\item{replace.backticks}{replace backticks (default: TRUE)}
}
\description{
Runs knit to convert from .Rmd to .md, as well as embed fonts.
}

