% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.panodc.R
\name{run.pandoc}
\alias{run.pandoc}
\title{Run pandoc with options}
\usage{
run.pandoc(file, output = stdout(), from = "markdown", to = "latex",
  options = "--chapters", pandoc.path = NULL)
}
\arguments{
\item{file}{input file for pandoc}

\item{output}{ouput file}

\item{to}{output file format, default "latex"}

\item{options}{string with options for pandoc, default "--chapters"}

\item{pandoc.path}{path to your pandoc executable, used if pandoc is not in your \code{PATH}}

\item{form}{input file format, default "markdown"}
}
\description{
Pandoc, written by John MacFarlane, converts files from one markup format into another. For more information see pandoc's website: \url{http://johnmacfarlane.net/pandoc/}.
}

