% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot as pdf to ./figure directory}
\usage{
save_plot(plot, figurename, width = get_plot_dim("width"),
  height = get_plot_dim("height"), caption = NULL, subcaption = NULL,
  tot.caption = NULL, notes = NULL, colormodel = "cmyk",
  extension = "pdf", path = "figure", fix.labels = FALSE,
  embed.fonts = TRUE, translate = TRUE, ...)
}
\arguments{
\item{plot}{plot object}

\item{figurename}{name of the figuer without extension}

\item{width}{width of the figure in cm}

\item{height}{height of the figure in cm}

\item{caption}{caption string, used in latex output}

\item{subcaption}{subcaption string, used in latex output}

\item{tot.caption}{caption string for the table of tables, used in latex output}

\item{notes}{notes that should be printed below the figure, used in latex output}

\item{colormodel}{colormodel of the pdf, default for print is \code{cmyk}}

\item{extension}{file extension, default is "pdf"}

\item{path}{path to the directory where the figure should be saved, relative to the manuscript. \code{../} is not allowed: TeX does not handle points in paths}

\item{fix.labels}{should the margins for the labels be adjusted? Used for \code{\link{directlabels}} with \code{last.bumpup}}

\item{embed.fonts}{embed fonts in pdf? Default is TRUE}

\item{translate}{logical, should the plot be translated before saving?}

\item{...}{arguments passed to \code{\link{pdf}}}
}
\description{
Save plot as pdf to ./figure directory and output latex code. This function depends on the format.plot package and will not function without it.
}

