% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_lang}
\alias{set_lang}
\title{Set Language and Dependent Variables}
\usage{
set_lang(lang = c("sv", "en"))
}
\arguments{
\item{lang}{language}
}
\description{
Set language and decimal mark, big mark, and global options for format.tables (read, write, render).
}
\seealso{
Other set-get-functions: \code{\link{get_lang}};
  \code{\link{get_year}}; \code{\link{set_year}}
}

