% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set-get-functions.R
\name{set_year}
\alias{set_year}
\title{Set year}
\usage{
set_year(year)
}
\arguments{
\item{year}{year}
}
\description{
Set year
}
\seealso{
Other set-get-functions: \code{\link{get_lang}};
  \code{\link{get_year}}; \code{\link{set_lang}}
}

