% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tbl_notes}
\alias{tbl_notes}
\title{Latex table notes}
\usage{
tbl_notes(tbl.notes = NULL, tbl.note.all = "", ...)
}
\description{
Latex table notes
}
\seealso{
Other latex: \code{\link{latex_table}};
  \code{\link{my.latexTranslate}};
  \code{\link{tbl_postamble}}; \code{\link{tbl_preamble}}
}

