% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex.r
\name{tbl_preamble}
\alias{tbl_preamble}
\title{Latex table preamble}
\usage{
tbl_preamble(style.indexes, data.col.count, caption = "", subcaption = "",
  tot.caption = "", first.col.width = "", col.type = "^r",
  cap.label.format = "empty", width = NULL, tbl.notes = FALSE,
  col.types.override = "", tbl.note.all = "", pos = "H", ...)
}
\description{
Latex table preamble
}
\seealso{
Other latex: \code{\link{latex_table}};
  \code{\link{my.latexTranslate}}; \code{\link{tbl_notes}};
  \code{\link{tbl_postamble}}
}

