% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colors.r
\name{rgb2cmyk}
\alias{cmyk2hex}
\alias{cmyk2rgb}
\alias{rgb2cmyk}
\title{Convert colors}
\usage{
rgb2cmyk(rgb, max = 255)

cmyk2rgb(cmyk, max = 100)

cmyk2hex(cmyk, alpha, max = 100)
}
\description{
Functions to convert between CMYK, RGB and HEX.
}
\seealso{
Other color functions: \code{\link{pm_colormap}},
  \code{\link{pm_colors}}, \code{\link{pm_colors}},
  \code{\link{pm_colors_cmyk}}, \code{\link{pm_colors_rgb}}
}

