% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/latex.r
\name{my.latexTranslate}
\alias{my.latexTranslate}
\title{Latex escape function}
\usage{
my.latexTranslate(my.string)
}
\arguments{
\item{my.string}{latex string}
}
\description{
Latex escape function
}
\seealso{
Other latex: \code{\link{latex_table}};
  \code{\link{tbl_notes}}; \code{\link{tbl_postamble}};
  \code{\link{tbl_preamble}}
}

