% Generated by roxygen2 (4.0.1): do not edit by hand
\name{output_table}
\alias{output_table}
\title{Read format.table, translate it, and output latex/html code}
\usage{
output_table(file, object.name = NULL)
}
\arguments{
\item{file}{the name of the file which the data are to be read from or an object of class 'format.tables'}

\item{object.name}{}
}
\description{
Read format.table, translate it, and output latex/html code
}

