% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-get-functions.R
\name{set_plot_dim}
\alias{get_plot_dim}
\alias{set_plot_dim}
\title{Set/get plot width and height}
\usage{
set_plot_dim(dim = c("width", "height"), value)

get_plot_dim(dim = c("width", "height"))
}
\arguments{
\item{dim}{dimension of the plot, either "width" or "height"}

\item{value}{numeric, value to set the dimension to}
}
\description{
Set/get plot width and height
}

