% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pm_format}
\alias{pm_format}
\title{Format numeric vectors}
\usage{
pm_format(x, trim = TRUE, scientific = FALSE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{further arguments passed to \code{\link{format}}}
}
\description{
Format numeric vectors with localized decimal and big marks
}

