% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/style-plot.r
\name{style_plot}
\alias{custom_theme}
\alias{style_plot}
\title{Style plot}
\usage{
style_plot(plot, colors = pm_colors(), base_size = 9)

custom_theme(base_size = 9, base_family = "Neo Sans Pro")
}
\description{
Functions to style plots
}

