% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/latex.r
\name{tbl_content}
\alias{tbl_content}
\title{Latex table content}
\usage{
tbl_content(data, style.indexes, tbl.notes, round.digits = 0,
  print.tbl.notes = TRUE, print.first.col.header = FALSE,
  print.header = TRUE, styles.override = NULL, ...)
}
\description{
Latex table content
}

