% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/latex.r
\name{tbl_postamble}
\alias{tbl_postamble}
\title{Latex table postamble}
\usage{
tbl_postamble()
}
\description{
Latex table postamble
}
\seealso{
Other latex: \code{\link{latex_table}};
  \code{\link{my.latexTranslate}}; \code{\link{tbl_notes}};
  \code{\link{tbl_preamble}}
}

